fix seal test output graph
in a previous refactor I noticed the QPen.setWidth(0.7) line and
corrected it to QPen.setWidthF to account for the floating point
value. the code actually relied on this value being truncated to 0.

this is now correctly documented in the code, next to other refactors
that allow one to actually read the code.

fixes #4